refactor: move test data to its own package
