+++ integration-tests/src/main/java/io/quarkiverse/operatorsdk/test/reconcilers/ChildTestResource.java
@@ -1,9 +1,10 @@
-package io.quarkiverse.operatorsdk.it;
+package io.quarkiverse.operatorsdk.test.reconcilers;
 
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
 
 @Group("example.com")
-@Version("v2")
-public class ChildTestResource2 extends TestResource {
+@Version("v1")
+public class ChildTestResource extends TestResource {
+
 }
