+++ integration-tests/src/main/java/io/quarkiverse/operatorsdk/test/reconcilers/ChildTestResource2.java
@@ -1,10 +1,9 @@
-package io.quarkiverse.operatorsdk.it;
+package io.quarkiverse.operatorsdk.test.reconcilers;
 
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
 
 @Group("example.com")
-@Version("v1")
-public class ChildTestResource extends TestResource {
-
+@Version("v2")
+public class ChildTestResource2 extends TestResource {
 }
