+++ integration-tests/src/main/java/io/quarkiverse/operatorsdk/test/reconcilers/SecretReconciler.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.operatorsdk.it;
+package io.quarkiverse.operatorsdk.test.reconcilers;
 
 import java.util.Base64;
 import java.util.HashMap;
